third_party/WebKit/LayoutTests/inspector/initial-modules-load.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

function test()
{
    InspectorTest.runTestSuite([
        function testInitialLoad(next)
        {
            InspectorTest.dumpLoadedModules(next);
        },

        function testCreateElementsPanel(next)
        {
            UI.inspectorView.panel("elements").then(InspectorTest.dumpLoadedModules.bind(InspectorTest, self.runtime.loadModulePromise("animation").then(next)));
        },

        function testCreateNetworkPanel(next)
        {
            UI.inspectorView.panel("network").then(InspectorTest.dumpLoadedModules.bind(InspectorTest, next));
        },

        function testShowSourcesPanel(next)
        {
            UI.inspectorView.panel("sources").then(InspectorTest.dumpLoadedModules.bind(InspectorTest, next));
        },

        function testOpenUISourceCode(next)
        {
            var resource;
            InspectorTest.resourceTreeModel.forAllResources(function(r) {
                if (r.url.indexOf("inspector-test.js") !== -1) {
                    resource = r;
                    return true;
                }
            });
            var uiLocation = Workspace.workspace.uiSourceCodeForURL(resource.url).uiLocation(2, 1);
            Common.Revealer.reveal(uiLocation);
            InspectorTest.dumpLoadedModules(next);
        }
    ]);
}

</script>
</head>

<body onload="runTest()">
<p>This test validates initial set of loaded modules.</p>
</body>
</html>
